Improving flake.nix: adding cache, activating flakes and nix-command, accept flake config; cleaning up Contributing + instruction to add trusted-user to the nix.config

+++ flake.nix
@@ -15,7 +15,7 @@
     # , cardano-node-flake
     , ...
     } @ inputs:
-    (flake-utils.lib.eachSystem [ "x86_64-darwin" "x86_64-linux" "aarch64-darwin" ]
+    (flake-utils.lib.eachSystem [ "x86_64-darwin" "x86_4-linux" "aarch64-darwin" ]
       (system:
       let
         pkgs = import nixpkgs {
@@ -77,7 +77,10 @@
                 # cardano-cli
               ];
               shellHook = ''
-                ln -s ${plutus}/plutus-conformance plutus-conformance
+                if [ ! -L plutus-conformance ]; then
+                  rm -rf plutus-conformance
+                  ln -s ${plutus}/plutus-conformance plutus-conformance
+                fi
                 echo "${pkgs.secp256k1}"
                 echo "${pkgs.libsodium}"
                 echo "${tiny_keccak_wrapper}"
@@ -120,10 +123,19 @@
   nixConfig = {
     extra-substituters = [
       "https://cache.iog.io"
+      "https://hydra.iohk.io"
+      "https://iohk.cachix.org"
+      "https://cache.nixos.org/"
+      "https://nix-community.cachix.org"
     ];
     extra-trusted-public-keys = [
       "hydra.iohk.io:f/Ea+s+dFdN+3Y/G+FDgSq+a5NEWhJGzdjvKNGv0/EQ="
+      "iohk.cachix.org-1:DpRUyj7h7V830dp/i6Nti+NEO2/nhblbov/8MW7Rqoo=" 
+      "cache.nixos.org-1:6NCHdD59X431o0gWypbMrAURkbJ16ZPMQFGspcDShjY="
+      "nix-community.cachix.org-1:mB9FSh9qf2dCimDSUo8Zy7bkq5CX+/rkCWyvRCYg3Fs="
     ];
     allow-import-from-derivation = true;
+    experimental-features = [ "nix-command" "flakes" ];
+    accept-flake-config = true;
   };
 }
